qt5-static: Fix CMake for static usage
Based on Stephen Kelly's reverted commit:

ae6d4c77eb338be3f2189bfaffdcd3c38cc57333
https://codereview.qt-project.org/#/c/33193/

which was reverted because INTERFACE_LINK_LIBRARIES
'expects a list of IMPORTED targets or full paths to
libraries'. As far as I can determine, passing linker
flags this way is a supported use-case (and AFAICT the
only way to pass linker flags). Further, it seems that
it works fine.

The CMake developers plan to un-tangle link libraries
from linker flags at some point:

http://public.kitware.com/pipermail/cmake-developers/2014-June/022281.html

and when that happens we should apply Stephen's revert
commit:

102e1822ffcdc9954d3c698f863734a8083e349c
https://codereview.qt-project.org/#/c/37307/

.. and following the suggestion in the commit message,
make changes to qmake so that it outputs the full path
to the libs in the prl files (QMAKE_PRL_LIBS_FOR_CMAKE)