git: Switch to a non-configure way to build it
It is actually easier and faster to build Git with the default
Makefile than to run configure first and possibly missing crucial
support (such as HTTP or Perl support).

Incidentally, it also makes the build script very small, and it
also makes it substantially easier to collaborate on the Git source
code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>