git: include pkgrel in pkgver, for continuously increasing versions
We need to be able to make multiple releases of Git for Windows from the
very same upstream version, in case things outside Git break, or only
Windows-specific bug fixes are needed.

In order not to confuse Pacman's version comparison, we therefore need a
way to let the version increase monotonically. If we only append the
short revision name, that is not the case: SHA-1s are pretty random,
practically speaking.

Unfortunately, the pkgrel needs to be appended *after* the pkgver,
cannot be empty, and must be numeric and consecutive.

So let's just insert the pkgrel into the version number before the short
revision name. It is slightly ugly to list the pkgrel twice in the file
name, but that cannot be helped. Time to move on.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>