git: build /git-bash.exe & /git-cmd.exe, too
In the MSys2 context, it makes most sense to run the Git Bash in a
mintty terminal emulator (it is a mere 170kB executable based on the
same msys2-runtime that the Bash already uses).

And while at it, also install the /cmd/ files: adding the `/cmd/`
directory to the `PATH` will make Git available without cluttering the
PATH with tons of `.dll` files that could possibly interfere with other
`.exe` files on the PATH.

Note: we have to be careful not to include symbols into the executables;
GCC wants to put symbols into them that are however ignored by Win32
API's resource update mechanism. Thus, after resource editing, the
symbols section is corrupted and in the best case, strip.exe (executed
as part of Pacman's makepkg) burps a "file truncated", in the worst case
it crashes with a segmentation fault.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>