Bump Git to 2.3.4 release 2
We actually never released v2.3.4.windows.1... but the tag exists, so
let's just go straight to release 2.

The difference between v2.3.4.windows.1 and v2.3.4.windows.2 is that the
Git wrapper now supports MSys2, too, and even allows us to compile a
`git-bash.exe` and a `git-cmd.exe` from the same source code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>