git: include the resources into `git-bash` the regular way
The attempt to teach the Git wrapper to be able to edit its command-line
resource was met with, let's put it mildly, objections.

Therefore this developer decided to not worry about making it possible
for, say, InnoSetup, or power users, to change the embedded command-line
themselves.

They will just have to use an interactive resource editor like ResEdit.

To allow the mingw-w64-git package to ship with the Git wrapper serving
as /cmd/git.exe, /git-bash.exe and /git-cmd.exe, we simply store the
respective resource files here and re-link the executables as part of
the building process.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>