mingw-w64-git: provide the option to parallelize git build
Since the maintainer relies on building this package with only one job
but others may want to use more than one lets give them the chance to do
so. If no `$PARALLEL_BUILD` is specified only one job is used. But if
needed `$PARALLEL_BUILD` can be used to provide the fully qualified
`-j<num>` parameter that will be passed to `make`. Example:

	PARALLEL_BUILD=-j5 makepkg-mingw -s

Note: `-j` without a numeric parameter caused hangs on the machine of
the developer.

Commit-Message-by: nalla <nalla@hamal.uberspace.de>
Signed-off-by: Kevin David <kevin@kevindavid.org>