git: clean up the version logic
Pacman has the concept of a `pkgver`, i.e. the version of a package, and
a `pkgrel`, i.e. a counter separate from the package version. This
allows the maintainer to make multiple releases from the same version,
e.g. when some software was packaged incorrectly.

Git for Windows has a very similar concept to the pkgrel: we build
multiple releases based on the same upstream version. For example, the
version v2.3.4 saw already three releases, tagged as v2.3.4.windows.1,
v2.3.4.windows.2 and v2.3.4.windows.3.

Earlier, we tried to interpret that Windows-specific release number as
identical to the pkgver, but that was wrong because the pkgrel changed,
but pkgver changed at the same time, which violates the pkgrel idea!

So let's try something new: let pkgrel do what it is supposed to do ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>