git: add /MINGW64/usr/share/git/builtins.txt
For the installer, it will be useful to know which files are builtins
(i.e. hardlinked versions of `git.exe` where hardlinks are supported).

We are in the best position to generate this list reliably when building
Git, so let's do that, and put it into a place where the installer (or
any other interested party) can pick it up easily.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>