git: generate documentation packages, too
For end users, the help pages are extremely valuable. So let's generate
them, too.

To make it a little bit more fine-grained, let's package the HTML pages
separately, as well as the man pages: with Git for Windows, we install
HTML pages (but not man pages because man(1) is not a native Windows
tool).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>