p11-kit: Use synchronous "spawn" instead of asynchronous "exec"
"exec" behaves like "spawn" with "_P_OVERLAY", i.e. it overlays a calling
process with a new process, destroying the calling process. This causes
problems if "p11-kit extract" is called in a script and the output file is
immediately used in the script as the internally called "trust" process
might not have finished writing the file by the time "p11-kit" returns
control to the script.