asciidoctor-extensions: allow installing 32/64-bit at the same time
In order to allow 32-bit and 64-bit MinGW packages at the same time,
MSys2 employs the strategy of encoding the architecture in the package
name itself and marking the package's architecture as 'any'. Otherwise,
Pacman would prevent i686 packages from being installed in x86_64 MSys2
and vice versa (which would make sense if the packages in question are
*MSys2* packages that differ from the architecture of the corresponding
msys2-runtime package).

We need to do the same here, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>