git-bash: ask for a new console to be allocated
This is required on XP where a recent change in the msys2-runtime
(actually, in Cygwin, on which the msys2-runtime is based) resulted in a
long delay before failing with a "Could not fork" message when `mintty`
was called from a GUI program. The culprit is that Cygwin expects either
no console at all, or a fully functional one, whereas on XP, a console
is provided that sports only a valid stdout, but no stdin nor stderr.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>