qtcreator: Fix issue 615
https://github.com/Alexpux/MINGW-packages/issues/615

Launching Qt Creator with MSYSTEM env. var set caused
Python to use / as os.path.sep which caused debugging
to fail due to an issue in the Python dumpers code.

It needs to replace all '\' with '/' in filenames but
instead replaced all os.path.sep with '/', which when
os.path.sep is already '/' was a noop.