git: adapt the git-wrapper to serve as bin/bash.exe, too
In Git for Windows 1.x, bash.exe and git.exe were in the /bin/
directory. For backwards-compatibility, we should have drop-in
executables in /bin/, too, that just hand off to the appropriate
executables. That job is exactly what the Git wrapper is for, so let's
adapt it to serve also as drop-in redirector for the Bash.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>