git: let Git Bash become properly pinned to the task bar
Windows 7 allows the user to pin running applications to the task bar.
However, since `git-bash.exe` actually launches `mintty.exe`, the latter
got launched instead of the former. In the meantime, we taught `mintty`
to offer better support for pinning to the task bar, and with this
change `Git Bash` uses that support.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>