Add the mingw-w64-connect package
The connect.exe helper is a very useful helper when trying to connect to
SSH through a SOCKS or HTTP proxy.

It comes in particularly handy when users are stuck behind restrictive
firewalls. If even one HTTP proxy is accessible that allows to connect to
port 22 of the desired outside machine, `connect.exe` can be used in
conjunction with the `ProxyCommand` directive in `$HOME/.ssh/config` to
allow regular ssh and scp connections. Example:

-- snippet of $HOME/.ssh/config --
Host github.com
User git
ProxyCommand connect -H proxy.local.net:8080 %h %p
-- end snippet

If a SOCKS proxy is available, the same directive works when `-H` is
replaced with the `-S` option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>