git: fix building without docs
The canonical way to restrict what `makepkg` (or `makepkg-mingw`) builds
is to pass the `--pkg=...` option. For example, if we want to build only
the 64-bit binaries, but not the documentation packages, we would call:

	makepkg-mingw --pkg=mingw-w64-x86_64-git

Of course, in that case we want to skip building the documentation
because it is not even packaged (and on Windows, it actually takes
substantially more time to build the documentation than the binaries,
thanks to xmlto slowness...).

The trick is to use the `PKGLIST` variable that `makepkg` sets
internally.

And we can finally get rid of the `GIT_SKIP_DOCS` hack that never worked
properly, anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>