Cleanup old environment if no ssh-agent process is found
Remove an old `SSH_AGENT_PID` environment variable if no ssh-agent
process is found. Having such a variable but no process can
happen if the user kills ssh-agent and restarted the agent
in an already opened session (cmd). The only "problem" was that
it printed a "Found ssh-agent at !SSH_AGENT_PID!" message which
actually was not true and did not remove the unused agent
socket files.

The startup happend anyway and afterwards the right env
variables were set, so this should be a cosmetic change.

Signed-off-by: Jan Schulz <jasc@gmx.net>