git: fix asciidoctor invocation
When creating the `config.mak` to tell `make` to use asciidoctor `instead`
of `asciidoc`, lets make sure that we only quote the parts that needs it.
Quoting the full `ASCIIDOCTOR_CONF` variable is leading to a wrong
invocation of the `asciidoctor` command because then the different
asciidoctor options `-I` and `-r` are not being recognized.

This fixes https://github.com/git-for-windows/git/issues/404

Signed-off-by: 마누엘 <nalla@hamal.uberspace.de>