Build the haru C-library for producing PDF files.
This library is support library for the pdf-related
drivers ad devices of the PLPLOT library.
This will bring the latest release version of libharu and build
it properly for both 64- and 32-bit mingw.  The small patch needed
for a 64-bit build was to use the (uintptr_t) cast for the IsAligned macro.

Tested by: Greg Jung <gvjung@gmail.com>
Testing of the libharu, or "hpdf", installation was undertaken in the
mingw64-msys2 context via a series of builds and example tests found
in plplot version 5.11.1.