git: deprecate PARALLEL_BUILD in favor of MAKEFLAGS
The `make` tool will interpret the environment variable `MAKEFLAGS`
before interpreting the command-line option, hence

	MAKEFLAGS=-j16 make

will ask `make` to use 16 processes in parallel.

This developer did not know that when accepting the patch to introduce
the `PARALLEL_BUILD` hack.

Let's deprecate that hack in favor of the tried-and-tested way to pass
options to `make` via an environment variable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>