curl: fix NTLM proxy when reusing connection
Under certain circumstances, an NTLM proxy connection will be reused.
If that is the case, the state of the authentication needs to be
handled properly.

There is a bug in cURL 7.44.0 that has been fixed in cURL's `master`
already (https://github.com/bagder/curl/commit/f65e07ca5) and this
patch backports it.

This fixes https://github.com/git-for-windows/git/issues/373

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>