gtk3: backport fixes for notebook tab-drag crashes
Fix some crashes seen when dragging tabs out of a notebook widget.
On the win32 backend, the second or later attempt to drag a tab out of
a GtkNotebook supporting reordering of tabs would segfault.

https://bugzilla.gnome.org/show_bug.cgi?id=756160

This commit backports the upstream fixes below onto GTK 3.18.2.
These will hopefully be incorporated into the release after 3.18.2,
so this commit can be reverted when MINGW-packages builds GTK >= 3.18.3.

1. 00b39e4a82401f407a2df2e2e35a18ede88b41ea
   "W32: Add missing reference taking in gdk_device_virtual_grab()"
2. c8604923598aad745a65ec16eb79113c97649345
   "Clean up handling of _gdk_win32_grab_cursor".
3. 25483304cdc967c374a435200416affb573583f8
   "Minor cleanup. Replace a unref-and-unset combination with g_clear_object."