SDL*_mixer: help ./configure find shared libraries for dynamic loading
This avoids hard dependencies against the libraries providing the
music backends (with the exception of libmad). Instead, if the
libraries are found in the search path, they are loaded as plug-ins at
run-time.

Fixes https://github.com/Alexpux/MINGW-packages/issues/865

First, the ./configure script did check only for $host being
*-mingw32*, so it skipped the relevant parts for $host=*-mingw64*.

Then, the ./configure script did check for the wrong library names for
both libfluidsynth*.dll and ibvorbisidec*.dll.

Finally, the find_lib() function in the ./configure script searched
for the shared libraries in the wrong directories. Fixed this by
adding $PATH to the search paths.