glib2: Fix pygtk/g_application_run/g_win32_get_command_line
When called from pygtk, the first argument to g_application_run
has been dropped from argv/argc (this is usually the name of
the python interpreter). However, in order to get argv as utf-8
on G_OS_WIN32 g_win32_get_command_line () is called, so this
dropping is undone.

To work around this, count the number of incoming arguments
and drop the same number of leading arguments as have newly
re-appeared after the call to g_win32_get_command_line ().

Fixes bug in meld3 where it would launch a 3-way compare
with the first pane containing meld3.py's source, and also
http://sourceforge.net/p/msys2/discussion/general/thread/e905716d/

Upstreaming at:

https://bugzilla.gnome.org/show_bug.cgi?id=734095