gcc: fix relocatex for 64-bit systems
This is a companion to a73c4fd (gettext: fix for 64-bit systems,
2016-02-04) and just like said commit, it ensures that the
win2unixpath() and canonicalize_file_name() functions are not assumed to
return a 32-bit int.

Please note that this developer was actually unable to get this code
compiled; src/build*/intl/ only contained the config.* files and the
Makefile, no object files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>