fontconfig: allow to disable auto font cache update
Each time fontconfig is installed or updated, a lengthy
process of font cache update is started automatically.
In certain use-cases, f.e. when one if its dependents
is used for automated build purposes, this causes an
unnecessary and unwanted delay in the process. On
AppVeyor CI free-tier in particular, this process takes
about 5 minutes.

This patch adds the ability to disable this automatic
font cache update, by setting the environment variable
`MSYS2_FC_CACHE_SKIP` to any non-empty value, prior
to lauching `pacman`.