mingw-w64-git: adjust git-bash to official mintty
Currently git-bash relies on mintty options that were proposed in a pull
request. A mintty package was also build with these changes. However, the
proposed options were changed slightly when merged into the official
mintty.

Offical mintty uses AppLaunchCmd instead of RelaunchCommand, AppName
instead of RelaunchDisplayName and requires an additional option
--store-taskbar-properties for the options to take effect.

In addition to these changes, specifying the path for the icon as
/mingw64/share/git/git-for-windows.ico no longer works. Using
"@@EXEPATH@@\\mingw64\\share\\git\\git-for-windows.ico" works, however,
instead of relying on a separate file we can just as well get the icon
from our git-bash executable.