git: move git-wrapper's sources into MINGW-packages
The original idea of the git-wrapper was to have a drop-in replacement
for msysGit's /cmd/git.cmd. Over the course of the years, it evolved to
be much, much more, including /git-bash.exe, /bin/bash.exe and also
serving as a drop-in redirector for the builtins.

One thing was a constant, though: the git-wrapper always was something
very specific to Git for Windows.

So let's pull it out of Git for Windows' fork of git.git (whose purpose
is to be a queue of patches that are either lined up to be submitted to
upstream Git or stabilizing to get there).

The source code reflects the state as of v2.9.3.windows.2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>