git: include the .ico again
We were too overzealous when we moved the .ico out of the mingw-w64-git
package: It is not only used for git-bash.exe, but was shipped so that
the .ico could be used more broadly.

We broke the icon for the Git entries in the Explorer integration, for
example.

Let's fix that regression by including the .ico in the package again.

This fixes https://github.com/git-for-windows/git/issues/870

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>