git: adjust to newer makepkg versions
Current makepkg versions do not store the list of packages to be built in
PKGLIST any longer: there is no longer a command-line option to limit
which packages are built.

Let's keep the clever code to build help packages only when needed,
though, so that editing the PKGBUILD to skip the -doc-html and  -doc-man
packages is easy: they simply need to be removed from the pkgname array.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>