Add new package ngspice-git.
The current stable ngspice package has known issues that cause problems
with KiCad during simulations.  The issues have been resolved in the
master branch of ngspice.  This package allows you to build the latest
development version of ngspice with the bug fixes.  Of course there is
no guarantee that new bugs have not been introduced.  Use with caution.