Remove rmtree debugging patch
It is shadowing exceptions thrown from the main "try" block of
GDumpParser._execute_binary_get_tree().

There does seem to be a race condition, but it seems to be related to
subprocess.check_call() and what it does on Windows. At the point
xml.[blah].parse() is called, "dump.xml" is often absent!