More tolerant rmtreeing
If shutil.rmtree() is failing due to some funky race condition that results
in files the walk doesn't know about being left, then don't worry about it
and continue.

This may leave uniquely named junk lying around in the build tree, but that
is by design not a problem with g-i-scanner. Just don't install the
tmp-introspectXXXXXX folder by accident.