Libarchive (#1936)
* mbedtls: Updated to 2.4.0 (#1868)

* mingw-w64-clang 3.9.0 - rev 2 delete files created by patch (#1867)

Delete files created by patches in case we are not deleting from srource
directory.

* jemalloc: Update to 4.2.1 (#1866)

* lua sqlite3 wrapper (#1861)

* Adds a TUI in GNU debugger, GDB. (#1869)

* nghttp2 - 1.16.0 Update to latest version (#1871)

Do not use --enable-lib-only for featureproofing.

* Readline (#1872)

* readline - 7.0 - Update to latest veersion - NOTE:  Will probably break stuff - rebuild what depends upon it

Rekeyed patch for new version and ifdef'ed out chown calls if not
available

* Rebuild some packages that depend upon readline for good measure

* libssh2 - 1.8.0 - Upgrade to latest version (#1876)

* Adding library: flickcurl (#1875)

* Adding library:flickcurl

* Flickcurl: fixing dependency on gtk-doc

* Libarchive (#1878)

* libarchive : 3.2.2

Update to altest version

* libarchive: 3.2.2 - add back dependency on lz4

* Nsis (#1879)

* nsis 3.0 update to release version

* mingw-w64-nsis: 33.

Download should point to main to main site, not mirror
Use https when possible

* g-i-parser: fix race condition (#1881)

* Remove rmtree debugging patch

It is shadowing exceptions thrown from the main "try" block of
GDumpParser._execute_binary_get_tree().

There does seem to be a race condition, but it seems to be related to
subprocess.check_call() and what it does on Windows. At the point
xml.[blah].parse() is called, "dump.xml" is often absent!

* More tolerant rmtreeing

If shutil.rmtree() is failing due to some funky race condition that results
in files the walk doesn't know about being left, then don't worry about it
and continue.

This may leave uniquely named junk lying around in the build tree, but that
is by design not a problem with g-i-scanner. Just don't install the
tmp-introspectXXXXXX folder by accident.

* giscanner debugging: asserts and waits

Make absence of dump.xml or functions.txt fatal, but only after giving them
plenty of time time to appear.

Yes, this is horrible code. But it fixes broken builds almost completely on
my 64-bit Win7 with NTFS.

* Wait for XML to be parsed completely

It's possible that dump.xml might exist but be incompletely written. This
nasty sleep-till-it-parses loop covers that case.

Plus the usual logging so we can see if it's really needed ☺

* gobject-introspection: bump pkgrel

* Updated to stable release 190600_20161030 (#1880)

A fresh new stable release since January 30, 2014!

* go: update to 1.7.3 (#1882)

* mingw-w64-libarchive - 3.2.2 - rev 3

Remove warning about a "cygpath -m /" reference pkg-config file.

* fix for bad formatting