git-wrapper: ensure that LC_ALL is defined
The LC_ALL variable is supposed to contain information about the locale
to use. If unset, it used to default to "C" but apparently recent MSYS2
versions switched to using "ASCII" as default (which makes little sense,
of course).

Let's just verify that LC_ALL is set, and fall back to setting LC_ALL=C
if it is not.

This fixes https://github.com/git-for-windows/git/issues/945

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>