Fix use of "git am": reinstate sensible versioning
This package was using plain "git am" to apply its patches, meaning that
revision counts were 2 revisions ahead of master, and that pkgver was
different every time because the hash was different.

This commit adds a temporary versioning scheme fix which will allow this
git package to get back on track without resorting to epochs. It can be
reverted later when it is not needed.