git: remove LC_ALL magic from the git-wrapper
The idea was to allow non-ASCII characters to be passed about freely.

However, in the meantime it was discovered that the only part of Git
for Windows that cannot handle non-ASCII characters without LC_* being
set appropriately is the POSIX emulation layer (also known as MSYS2
runtime).

Meaning that the git-wrapper is the wrong place to fix this: instead,
we just changed `git.exe` to ensure that LC_* is set. This way, the
git-wrapper can still call Git Bash (which initializes the locales
appropriately itself and would suffer from an incorrectly pre-set
LC_ALL or LC_CTYPE).

This fixes https://github.com/git-for-windows/git/issues/1036

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>