mingw-w64-curl: new version
This required a backport of a fix that already made it into cURL's
master branch for a build error discovered only after 7.53.0 was
released.

While at it, I just regenerated the patch files for easier future
adjustments (of course, the easiest would be to just maintain these
patches in a long-running Git branch instead of patch files, but that
is too much to hope for).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>