mingw-w64-git: use `make` to build all of the git-wrapper stuff
Previously, we built the git-wrapper.exe via `make` and then compiled
stuff ourselves using direct calls to `windres` and `gcc`.

This was a historical left-over from the time where Git for Windows'
for of Git's source code had a Makefile target for `git-wrapper.exe`,
and we had add-on stuff in MINGW-packages/mingw-w64-git/ for the
derivatives (such as `cmd/git.exe`).

Seeing as we already use `make` to build the `git-wrapper.exe`, and now
write out a `git-wrapper.mak` file containing the Makefile rules, we
might just as well extend it to build the rest, too.

As a bonus, we no longer have to duplicate the logic for the ASLR flags,
as we simply include Git's Makefile and inherit all settings.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>