Add libwebp as runtime dependency to opencv.
The current binary package
"mingw-w64-x86_64-opencv-3.2.0-3-any.pkg.tar.xz"
links to the outdated DLL "libwebp-6.dll" but doesn't
specify libwebp as a dependency.

Since libwebp is updated in the meanwhile to "libwebp-7.dll",
it isn't sufficient to install the package.

Rebuilding opencv with webp dependency fixes this issue.