cmake and some updated dependencies (#2633)
* cmake and some updated dependencies
mingw-w64-cmake - 3.8.2 - Update to latest release version and remove 0002-bootstrap-Set-DEFAULT_CODEPAGE-in-cmake_cxx_flags_En.patch .  That patch is not needed for bootstrap
mingw-w64-expat - 2.2.1 - Update to latest version, add testing as part of build process.  Add 002-fix-tests.patch so those can work in MSYS environment.  remove expat-2.2.0-CVE-2016-0718-regression.patch as that is in the source-code.
mingw-w64-lncruses - 6.0 - 20170624 - Update to latest version

* Fix various PKGBUILD errors (#2631)

These PKGBUILD files could no longer be openend by makepkg due to
stricter parsing in regard to variable types. This fixes the types.

* libpeas: Python 3 rebuild (#2627)

* mbedtls: Updated to 2.5.1 (#2629)

* python-sphinx: Fix dependencies (#2628)

* jasper: Added patch file to fix static library link error. (#2623)

To use the static library you need to define JAS_DLL to zero.
-D JAS_DLL=0

* expat: Update to 2.2.1 (#2625)

Includes various security fixes.
Drop patch applied upstream.

* Armadillo: update to 7.950.1 (#2624)

Fixes #2618

* glib2: Update to 2.52.3 (#2621)

* nodejs: update to 6.11.0 and include npm (#2622)

* OsgEarth and gdal improvments (#2620)

* osgearth : active osgEarthQt module

* gdal : change PROJ4 library name gdal try to dlopen libproj-9.dll but correct name is libproj-12.dll

* python3: Make multiprocessing.Queue work; Add some tests. Fixes #837 (#2619)

With non-mingw builds POSIX_SEMAPHORES_NOT_ENABLED is not enabled and
sem_unlink is a no-op. Change it to always be provided on Windows instead.

This adds a minimal test suite which makes it easier to catch regressions
in the future for common MSYS2 related bugs.

* vulkan: add sdk (#2590)

* glslang: update for vulkan sdk

* vulkan: add sdk
*vulkan-headers are included in sdk
*vulkan-docs are separate package because of versions difference

* vulcan: Fix PKGBUILD. Add conflicts

* Update PKGBUILD

Attempt to fix this build.

* Update checksums