git: backport patches to v2.13.1(2) for a new MinGit-only release
One of the most prominent consumers of MinGit is Visual Studio. To
benefit from the important bug fix in v2.14.1 and from the improved
performance introduced by the --show-ignored-directory flag, these
patches were backported so that a future Visual Studio update can ship
with a minimal update including MinGit v2.13.1(3).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>