git: also package test artifacts
The test artifacts can be used to test a BusyBox-backed MinGit in-place,
i.e. without a Git for Windows SDK.

Note: this is currently not completely true, as some parts of the test
suite still rely on the presence of a Perl interpreter (even when
NO_PERL is set). The hope is that we will be able to fix this quickly
after identifying the tests which will fail when removing /usr/bin/perl
from the MinGit with test artifacts.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>