git (git-gui): prevent wish from interpreting git-gui options
While the /cmd/git-gui.exe wrapper was never intended to be used
directly via the command-line (`git-gui` is the long-deprecated dashed
form, after all), it *should* ideally be possible to pass command-line
parameters to that wrapper.

One reported problem was that wish.exe thinks that the options starting
with a dash are intended for its use. What we need to do is simply
separate the git-gui options from the (non-existing) wish options using
a double dash.

This fixes https://github.com/git-for-windows/git/issues/1284

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>