SDL2_Mixer: link in MP3, Ogg Vorbis and FLAC support libraries (#3114)
* link in MP3, Ogg Vorbis and FLAC support libraries

Link in the libraries instead of relying on dlopen().

Support for these formats is pretty much expected nowadays and we
depend on the corresponding packages anyway. I am not applying this
to modplug and fluidsynth as well, because they support more exotic
formats and pull in some rather heavy-weight dependencies
(e.g. libglib in the case of fluidsynth).

* extend SDL2_mixer-2.0.1-find_lib.mingw.patch to libmpg123