git: don't tell GCC to minimize the size for the Git wrapper
We ship this git-wrapper that essentially just sets a couple of
environment variables and then executes a different executable with the
command-line parameters handed down to the child process.

This git-wrapper is so tiny because we want to use it for Git's
builtins, for cmd\*.exe and for bin\bash.exe. And in earlier tests,
compiling normally and then calling strip.exe would not result in as
small binaries as using the -s flag right away.

However, we will change the way the binaries are "stripped": in the next
commit, we will start to use cv2pdb to extract the symbols into .pdb
files, and the resulting .exe file is comparably small to the one
produced by linking with -s. And cv2pdb is really unhappy if it finds no
symbols (which is the case when linking with -s).

So let's just not link with -s anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>