openssl: process correct *eay32.dll via cv2pdb
The *eay32.dll files that actually get bundled are not the ones in
apps/, but in the top-level directory.

As a consequence, this maintainer now built -2, -3 & -4 versions of
OpenSSL 1.0.2n with *huge* libeay32.dll (in the 64-bit version, we are
talking about 2.3MB -> 15MB).

Let's process the correct files instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>