cmake: Remove dead unit-test munging
This code appears to have been updating the paths in a CMake Unit Test
to match the live paths in MSYS2.

However, it was disabled when added (see #2633), and doesn't make sense
since the test is only of the parsing of the output format, not the
contents of the output.