python3: include official .rc files for the dlls/execs. See #3139
Use windres to build the resource files and include them in the
dll and executables. Since we need a different resource for the
console/gui version move the creation of the python3w.exe file from
the PKGBUILD into the Makefile.

With a bit of luck this should allow one to opt into the new long paths
feature on Windows 10, like with the official builds.
(no Windows 10 here, can't test)