curl: prepare the "make cURL relocatable" patch for v7.58.0
This is now a straight `format-patch` artifact produced from the tip of
https://github.com/dscho/curl/tree/relocatable-curl-7.58.0.

Note: this branch is *not* based on the regular curl-7_58_0 tag, as we
want to patch the source code as unpacked from curl-7.58.0.tar.bz2
(which has autoconf-generated files in it that we need to patch, too).

For good measure, the branch corresponding to the previous patch is
here: https://github.com/dscho/curl/tree/relocatable-curl-7.57.0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>