meson: use sys.prefix as default prefix and strip the prefix when installing with DESTDIR (#3324)
Previously the prefix was just C:\ which made install work as with autotools, but
was obviously wrong. meson also allows passing absolute install paths which
ignore the prefix and then result in <destdir>/msys64/mingw32/<sub> paths.

To fix this set the right prefix by default and after the install path is computed
strip away the prefix again.