gdk-pixbuf2: build all available loaders and only use gdi loaders where needed. See #3335
gdk-pixbuf only has a mode for building gdi loaders and then skipping
all formats supported through gdi, and not building gdi and building
all formats.

Since the gdi loaders can be slow, but we still want to get the extra format
support not covered by other loaders build all loaders including gdi
and the just don't register the gdi loaders we don't want to use.

As it's always a pain to test if all the loaders are OK and the right ones
are used this adds a Python script which tests each loader with some
example image data and also makes sure if the gdi one is used or not.

In addition this enables the jpeg2000 loader using libjasper
and backports a patch from upstream master to make the libtiff loader
work when loading from filenames.