Several other testing things that need to reference the proper interpretter and an updated beaker.  These should go together because tof dependencies.
Note that it does seem necessary to have a path reference to the proper interpretter instead of using env because it might reference the wrong classes from the wrong python library.  That can cause problems if the MSYS version is different than the ones for mingw32 and mingw64.  Thus, we should do a fixup in the *.install scripts like python-pip.

beaker - update to version 1.9.0 and rework to conform with the pattern I use.
make - make sure the command line .exe script uses the right interpreter and classes.
nose - make sure the command line .exe script uses the right interpretter and classes