Add files via upload
This pull request takes the latest netcdf version (4.6.0, previous was 4.5.0) and ADDS OPENDAP functionality.
I am currently discussing with the netcdf people if some patches can be permanently in the source. No promises yet.

Note this requires a slightly modified version of the curl library to work properly. I will issue a pull request for that too